Sorry mate

+++ features/src/main/kotlin/ru/mobileup/template/features/crypto/ui/coins_list/list/CoinItem.kt
@@ -74,7 +74,7 @@ fun CoinItem(
             Text(
                 text = getFormattedPercentage(coin.priceChangePercentage),
                 style = CoinTheme.typography.coin.normal,
-                color = if (coin.priceChangePercentage.toDouble() < 0)
+                color = if (coin.priceChangePercentage < 0)
                     CoinTheme.colors.text.percentBelowZero
                 else
                     CoinTheme.colors.text.percentAboveZero
